Added snippet object.
Necessary to access elements of a Sublime Text snippet. Sublime Text
snippets are in XML format so this object uses `REXML` to read the XML
snippet and retrieve the elements of the snippet. Access is done via
metaprogramming since behavior is the same for each element of the
snippet. Example XML:

    <snippet>
      <content><![CDATA[reduce(0, :+)]]></content>
      <tabTrigger>sum</tabTrigger>
      <description>Ruby Sum</description>
      <scope>source.ruby, source.rails</scope>
    </snippet>

Using the above snippet, we'd be able to get the content of each:

    snippet = SublimeTextKit::Snippets::Snippet.new document
    snippet.content     # => "reduce(0, :+)"
    snippet.trigger     # => "sum"
    snippet.description # => "Ruby Sum"
    snippet.scope       # => "source.ruby, source.rails"