Refactored session as a session rebuilder
Lots of changes were applied in this refactor to where this object
adheres to the *Command Pattern* now. This also means the object is
much easier to construct with safe defaults. The metadata dir is now an
argument that can be passed to the `#call` method. All methods were
converted to endless methods and Pathnames were used where possible to
reduce complexity further.